Converted a number of ints to size_t in SAPT code (#619)
Should prevent integer overflow. May not be exhaustive though.

+++ psi4/src/psi4/libsapt_solver/sapt.h
@@ -64,19 +64,19 @@ protected:
   std::shared_ptr<BasisSet> elstbasis_;
   std::shared_ptr<BasisSet> zero_;
 
-  int nsoA_;
-  int nmoA_;
-  int nsoB_;
-  int nmoB_;
-  int ndf_;
-  int noccA_;
-  int foccA_;
-  int aoccA_;
-  int noccB_;
-  int foccB_;
-  int aoccB_;
-  int nvirA_;
-  int nvirB_;
+  size_t nsoA_;
+  size_t nmoA_;
+  size_t nsoB_;
+  size_t nmoB_;
+  size_t ndf_;
+  size_t noccA_;
+  size_t foccA_;
+  size_t aoccA_;
+  size_t noccB_;
+  size_t foccB_;
+  size_t aoccB_;
+  size_t nvirA_;
+  size_t nvirB_;
   int NA_;
   int NB_;
   int natomsA_;
@@ -110,7 +110,7 @@ protected:
 
   std::shared_ptr<SAPTDenominator> denom_;
 
-  int nvec_;
+  size_t nvec_;
 
   double **dAR_;
   double **dBS_;
@@ -130,7 +130,7 @@ class CPHFDIIS {
 
 private:
   int max_diis_vecs_;
-  int vec_length_;
+  size_t vec_length_;
 
   int curr_vec_;
   int num_vecs_;
